+++ changelog.md
@@ -1,3 +1,13 @@
+[4.2.0 - 2024-06-11]
+- Updating of `parseResults`, changing the order of 'for's: for each command, all templates are tested in `match = prog.search(cmdsLogs)`. In this way, the existence of command log determines the presence of the data in the excel sheet. In previous versions we used the reverse order, so to have data in excel, it was necessary to have the specific template to parse.
+- New: Generic template.
+  - If a command/log don't match with any template available in the specified folder with templates, the parsing is done with `general.template`, available in src. Add structure in `readTemplate`, `parseResults`;
+  - As many sheets as needed can be created by generic template;
+- `parseResults` with new functions `detParseStatus` and `writeDfTemp`.
+- `datosEquipo[tmpltName]['template']` in `datosEquipo[tmpltName]`, to better identification in cases with generic template, when this key is related to `dTmplt`. For example, in `searchDiffAll`, `searchDiffOnly`, `findMajor`.
+- Index: for better identification, column `Command` added. 
+
+
 [4.1.0 - 2024-04-10]
 
 - New structure for `datosEquipo[tmpltName]`. Now with `datosEquipo[tmpltName]['parseStatus']` and `datosEquipo[tmpltName]['dfResultDatos']`. Created for better detection of the new statuses of parsing. They specify "no parsing" cases;